Create a New Pokémon

R6 object containing required information about a Pokémon.

Methods

Public methods

Pokemon$new()

Pokemon$status()

Pokemon$use_move()

Pokemon$get_stat()

Pokemon$get_moves()

Pokemon$change_move()

Pokemon$get_crit_chance()

Pokemon$take_damage()

Pokemon$clone()

Method new()

Create a Pokémon

Usage

Pokemon$new(
  pokemon = NULL,
  level = 50L,
  generation = 1L,
  nature = NULL,
  api_data = NULL,
  language = "en"
)

Arguments

pokemon

Name of the Pokémon to create

level

The level the Pokémon should be

generation

The generation that the Pokémon comes from

nature

Either a specified nature, or NULL for a random nature. NB Nature is only required from generation 3 onwards

api_data

Either data pulled from pokeapi::get_pokemon, or NULL to select the data of a random Pokémon

language

ISO-2 character of the language to pull the name of the Pokémon

Returns

A Pokémon

Method status()

Show current status of the Pokémon

Usage

Pokemon$status(simple = FALSE, console = TRUE)

Arguments

simple

Logical, do you just want the simple status (name + HP) printed?

console

Logical, should the status be printed or returned as a string?

Method use_move()

Use an attack of the Pokémon

Usage

Pokemon$use_move(move, def_pokemon, battle)

Arguments

move

The name of the move that is being used

def_pokemon

The defending Pokemon

battle

A PokemonBattle

Method get_stat()

Get the stat of the Pokémon

Usage

Pokemon$get_stat(stat)

Arguments

stat

The private field of the Pokémon

Method get_moves()

Get the moveset of the Pokémon

Usage

Pokemon$get_moves()

Method change_move()

Change one of the moves of the Pokémon

Usage

Pokemon$change_move(new_move = NULL, replace_move = NULL)

Arguments

new_move

Name of the new move

replace_move

Either the name of the move to replace, or the position of the move

Method get_crit_chance()

Get the critical hit chance for a move used by the Pokémon

Usage

Pokemon$get_crit_chance(move)

Arguments

move

Name of the move used by the Pokémon

Method take_damage()

Take damage from attack

Usage

Pokemon$take_damage(damage_dealt)

Arguments

damage_dealt

The amount of damage dealt by the opposing Pokémon's attack

Method clone()

The objects of this class are cloneable with this method.

Usage

Pokemon$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.